+++ src/aux-common/partitions/MemoryPartition.ts
@@ -221,7 +221,7 @@ export class MemoryPartitionImpl implements MemoryPartition {
                         const newVal = event.update.tags[tag];
                         const oldVal = newBot.tags[tag];
 
-                        if (newVal !== oldVal) {
+                        if (newVal !== oldVal || Array.isArray(newVal)) {
                             changedTags.push(tag);
                         }